+++ core/src/main/scala/slinky/core/BaseComponentWrapper.scala
@@ -13,16 +13,28 @@ import scala.reflect.macros.whitebox
 
 final class KeyAndRefAddingStage[D](private val args: js.Array[js.Any]) extends AnyVal {
   @inline def withKey(key: String): KeyAndRefAddingStage[D] = {
+    if (args(0) == null) {
+      throw new IllegalStateException("This component has already been built into a ReactElement, and cannot be reused")
+    }
+
     args(1).asInstanceOf[js.Dictionary[js.Any]]("key") = key
     this
   }
 
   @inline def withRef(ref: D => Unit): KeyAndRefAddingStage[D] = {
+    if (args(0) == null) {
+      throw new IllegalStateException("This component has already been built into a ReactElement, and cannot be reused")
+    }
+
     args(1).asInstanceOf[js.Dictionary[js.Any]]("ref") = ref
     this
   }
 
   @inline def withRef(ref: ReactRef[D]): KeyAndRefAddingStage[D] = {
+    if (args(0) == null) {
+      throw new IllegalStateException("This component has already been built into a ReactElement, and cannot be reused")
+    }
+
     args(1).asInstanceOf[js.Dictionary[js.Any]]("ref") = ref
     this
   }
@@ -30,8 +42,16 @@ final class KeyAndRefAddingStage[D](private val args: js.Array[js.Any]) extends 
 
 object KeyAndRefAddingStage {
   @inline implicit def build[D](stage: KeyAndRefAddingStage[D]): ReactElement = {
-    ReactRaw.createElement
+    if (stage.args(0) == null) {
+      throw new IllegalStateException("This component has already been built into a ReactElement, and cannot be reused")
+    }
+
+    val ret = ReactRaw.createElement
       .applyDynamic("apply")(ReactRaw, stage.args).asInstanceOf[ReactElement]
+
+    stage.args(0) = null
+
+    ret
   }
 }